feature : 급여 입력 오류수정중

+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/hr/repository/basic_information_management/salary_ledger/IncomeTaxRepository.java
@@ -3,6 +3,6 @@ package com.megazone.ERPSystem_phase2_Backend.hr.repository.basic_information_ma
 import com.megazone.ERPSystem_phase2_Backend.hr.model.basic_information_management.salary.IncomeTax;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IncomeTaxRepository extends JpaRepository<IncomeTax, Integer> {
+public interface IncomeTaxRepository extends JpaRepository<IncomeTax, Long> {
 
 }
